+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project are documented in this file.
 ### Added
 
 - GM Sidebar now supports scrollable inventory and stats.
+- Connection presets! This allows for simpler setup of API connections and switching back and forth between different profiles.
+- Instructions on how to setup each provider.
 
 ### Fixed